+++ dalesbred/src/main/java/org/dalesbred/annotation/DalesbredInstantiator.java
@@ -26,16 +26,16 @@ import java.lang.annotation.*;
 
 /**
  * <p>
- * Marks a constructor as Dalesbred instantiator. This means that when Dalesbred tries to instantiate
- * classes of this type, it skips the normal lookup resolution and will always use this constructor.
+ * Marks a constructor or a static method as Dalesbred instantiator. This means that when Dalesbred tries to
+ * instantiate classes of this type, it skips the normal lookup resolution and will always use this constructor.
  * It will not search for other constructors, nor will it set any properties.
  * </p>
  * <p>
- * It is an error to mark multiple constructors as instantiators.
+ * It is an error to mark multiple constructors/methods as instantiators.
  * </p>
  */
 @Retention(RetentionPolicy.RUNTIME)
 @Documented
-@Target(ElementType.CONSTRUCTOR)
+@Target({ElementType.CONSTRUCTOR, ElementType.METHOD})
 public @interface DalesbredInstantiator {
 }