+++ dalesbred/src/main/java/org/dalesbred/internal/instantiation/ReflectionInstantiator.java
@@ -26,6 +26,8 @@ import org.dalesbred.internal.utils.Throwables;
 import org.jetbrains.annotations.NotNull;
 
 import java.lang.reflect.Constructor;
+import java.lang.reflect.Executable;
+import java.lang.reflect.Method;
 import java.util.List;
 
 import static java.util.Objects.requireNonNull;
@@ -35,27 +37,39 @@ import static java.util.Objects.requireNonNull;
  */
 final class ReflectionInstantiator<T> implements Instantiator<T> {
 
-    private final @NotNull Constructor<T> constructor;
-
+    private final @NotNull Executable instantiator;
     private final @NotNull List<TypeConversion> conversions;
 
     private final @NotNull List<PropertyAccessor> accessors;
 
-    private final int constructorParameterCount;
+    private final int parameterCount;
 
-    ReflectionInstantiator(@NotNull Constructor<T> constructor,
+    ReflectionInstantiator(@NotNull Executable instantiator,
                            @NotNull List<TypeConversion> conversions,
                            @NotNull List<PropertyAccessor> accessors) {
-        this.constructor = requireNonNull(constructor);
+        this.instantiator = requireNonNull(instantiator);
         this.conversions = requireNonNull(conversions);
         this.accessors = requireNonNull(accessors);
-        this.constructorParameterCount = constructor.getParameterTypes().length;
+        this.parameterCount = instantiator.getParameterTypes().length;
     }
+
 
     @Override
     public @NotNull T instantiate(@NotNull InstantiatorArguments arguments) {
         try {
-            T value = constructor.newInstance(constructorArguments(arguments.getValues()));
+            Object[] argumentArray = toArgumentArray(arguments.getValues());
+
+            Object v;
+            if (instantiator instanceof Constructor<?>) {
+                v = ((Constructor<?>) instantiator).newInstance(argumentArray);
+            } else if (instantiator instanceof  Method) {
+                v = ((Method) instantiator).invoke(null, argumentArray);
+            } else {
+                throw new IllegalStateException("Unexpected instantiator: " + instantiator);
+            }
+
+            @SuppressWarnings("unchecked")
+            T value = (T) v;
             bindRemainingProperties(value, arguments);
             return value;
         } catch (Exception e) {
@@ -67,15 +81,15 @@ final class ReflectionInstantiator<T> implements Instantiator<T> {
         List<?> values = arguments.getValues();
 
         for (int i = 0, len = accessors.size(); i < len; i++) {
-            int argumentIndex = i + constructorParameterCount;
+            int argumentIndex = i + parameterCount;
             Object originalValue = values.get(argumentIndex);
             Object convertedValue = conversions.get(argumentIndex).convert(originalValue);
             accessors.get(i).set(result, convertedValue);
         }
     }
 
-    private @NotNull Object[] constructorArguments(@NotNull List<?> arguments) {
-        Object[] result = new Object[constructorParameterCount];
+    private @NotNull Object[] toArgumentArray(@NotNull List<?> arguments) {
+        Object[] result = new Object[parameterCount];
 
         for (int i = 0; i < result.length; i++)
             result[i] = conversions.get(i).convert(arguments.get(i));