+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ### Changes
 
+  - Support @DalesbredInstantiator on static methods
   - Migrate build to Gradle 7.2
   - Specify optional dependencies as Gradle's features
   - Bump dependencies to latest versions